+++ app/markets/tests/test_admin.py
@@ -1,98 +1,43 @@
-from django.contrib.auth.models import Permission, User
-from django.core.urlresolvers import reverse
-from django.test import TestCase
+from django.http import HttpResponse
 
-from ..models import Market, PublishedMarket
-from . import create_market, get_market_data
+import pytest
+from django.urls import reverse
+from django.test.client import Client
+from django.contrib.auth.models import User
 
 
-class MarketAdminTests(TestCase):
+@pytest.fixture
+def response_with_session_cookie(settings):
+    settings.SESSION_COOKIE_NAME = 'session'
+    response = HttpResponse()
+    response.set_cookie(
+        settings.SESSION_COOKIE_NAME,
+        value='123',
+        domain=settings.SESSION_COOKIE_DOMAIN,
+        max_age=settings.SESSION_COOKIE_AGE,
+    )
+    return response
 
-    @classmethod
-    def setUpClass(cls):
-        cls.username = 'testuser'
-        cls.password = '12345'
-        cls.user = User.objects.create(username=cls.username)
-        cls.user.set_password('12345')
-        cls.user.is_staff = True
-        cls.user.save()
-        cls.change_perm = Permission.objects.get(codename='change_market')
-        cls.add_perm = Permission.objects.get(codename='add_market')
-        cls.publish_perm = Permission.objects.get(codename='can_publish')
 
-    @classmethod
-    def tearDownClass(cls):
-        cls.user.delete()
+@pytest.fixture(autouse=False)
+def admin_user():
+    admin_user = User.objects.create_superuser('myuser', 'myemail@test.com', 'password')
+    admin_user.is_staff = True
+    admin_user.save()
+    return admin_user
 
-    def _add_permission(self, permission):
-        self.user.user_permissions.add(permission)
-        self.user = User.objects.get(username=self.username)
-        self.client.logout()
-        self.client.login(username=self.user.username, password=self.password)
 
-    def setUp(self):
-        self.client.login(username=self.user.username, password=self.password)
+@pytest.fixture()
+def admin_client_sso(db, admin_user):
+    """A Django test client logged in as an admin user."""
+    client = Client()
+    client.login(username=admin_user.email, password=admin_user._password)
+    return client
 
-    def tearDown(self):
-        self.user.user_permissions.clear()
-        self.client.logout()
 
-    def test_add_no_publish(self):
-        self._add_permission(self.add_perm)
-        self._add_permission(self.change_perm)
-        self.assertEqual(Market.objects.count(), 0)
+@pytest.mark.django_db
+def test_admin_permission_middleware_authorised_with_staff(client, settings, admin_user):
+    client.force_login(admin_user)
+    response = client.get(reverse('authbroker_client:login'))
 
-        market_data = get_market_data()
-        response = self.client.post(reverse('admin:markets_market_add'),
-                                    market_data)
-
-        self.assertRedirects(response,
-                             reverse('admin:markets_market_changelist'))
-        self.assertEqual(Market.objects.count(), 1)
-        self.assertEqual(PublishedMarket.objects.count(), 0)
-
-    def test_publish_button(self):
-        market = create_market()
-        self._add_permission(self.publish_perm)
-        self._add_permission(self.change_perm)
-        response = self.client.get(
-            reverse('admin:markets_market_change', args=[market.pk]))
-        self.assertContains(response, 'name="_publish"')
-
-    def test_no_publish_button(self):
-        market = create_market()
-        self._add_permission(self.change_perm)
-        response = self.client.get(
-            reverse('admin:markets_market_change', args=[market.pk]))
-        self.assertNotContains(response, 'name="_publish"')
-
-    def test_publish(self):
-        market = create_market()
-        self._add_permission(self.publish_perm)
-        self._add_permission(self.change_perm)
-
-        self.assertEqual(PublishedMarket.objects.count(), 0)
-
-        # Try and publish, but validation will mean it won't publish
-        response = self.client.post(
-            reverse('admin:publish_market', args=[market.pk]))
-        self.assertRedirects(response, reverse('admin:markets_market_change',
-                                               args=[market.pk]))
-        self.assertEqual(Market.objects.count(), 1)
-        # XXX: CAn I test the redirected page contains
-        # "Failed to publish Market" as a user message?
-        self.assertEqual(PublishedMarket.objects.count(), 0)
-
-    def test_admin_restricted(self):
-        with self.settings(RESTRICT_ADMIN=True):
-            self._add_permission(self.add_perm)
-            self._add_permission(self.change_perm)
-            self.assertEqual(Market.objects.count(), 0)
-
-            market_data = get_market_data()
-            response = self.client.post(
-                reverse('admin:markets_market_add'),
-                market_data,
-                **{'HTTP_X_FORWARDED_FOR': '74.125.224.72'}
-            )
-            self.assertEqual(response.status_code, 404)
+    assert response.status_code == 302