+++ app/navigator/urls.py
@@ -12,6 +12,8 @@ urlpatterns_unprefixed = [
     url(r'^robots\.txt$', TemplateView.as_view(template_name='robots.txt', content_type='text/plain'), name='robots'),
     url(r'^ping\.json$', PingView.as_view(), name='ping'),
     url(r'^grappelli/', include('grappelli.urls')),
+    url('^auth/', include('authbroker_client.urls', namespace='authbroker', app_name='authbroker_client')),
+    url(r'^admin/login/$', RedirectView.as_view(url=reverse_lazy('authbroker:login'), query_string=True)),
     url(r'^admin/', admin.site.urls),
     url(r'^$', HomepageView.as_view(), name='home'),
     url(r'^markets/', include('markets.urls'), name="markets"),
@@ -19,12 +21,6 @@ urlpatterns_unprefixed = [
     url(r'^geography/', include('geography.urls'), name="geography"),
     url(r'^activity-stream/v1/', ActivityStreamViewSet.as_view({'get': 'list'}), name='activity-stream'),
 ]
-
-if settings.FEATURE_FLAGS['ENFORCE_STAFF_SSO_ON']:
-    urlpatterns_unprefixed = [
-        url('^auth/', include('authbroker_client.urls', namespace='authbroker', app_name='authbroker_client')),
-        url(r'^admin/login/$', RedirectView.as_view(url=reverse_lazy('authbroker:login'), query_string=True)),
-    ] + urlpatterns_unprefixed
 
 # to display thumbnails properly MEDIA_URL needs to have added prefix separately
 urlpatterns = [