+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Pre-release
 - no ticket - update components for new cookie banner
 - MVP-40 - Add GA360 context processor and remove GA360 Middleware
+- Remove existing SSO Middleware and move to directory-components SSO Middleware
+- Remove feature flagging for Authbroker Staff SSO
+- Remove unneccessary code in Market View
+- Fix Market homepage bug showing unpublished markets to logged-in users
+- Upgrade directory-components to 35.17.0
 
 
 ## [2019.02.04] (https://github.com/uktrade/navigator/releases/tag/2019.02.04)